timers seem to be fixed. Stable

+++ src/ru/narod/vn91/pointsop/gui/RoomPart_Userlist.java
@@ -76,7 +76,7 @@ public class RoomPart_Userlist extends javax.swing.JPanel {
 			playerNew.addChangeListener(new PlayerChangeListener() {
 				@Override
 				public void onChange(Player player) {
-					if (player.rating!= playerRatingBeforeChange.value) {
+					if ((int)player.rating!= (int)playerRatingBeforeChange.value) {
 						System.out.println("hard user rejoin executed: " +
 							player + " " +
 							playerRatingBeforeChange.value +
